refactor: 테이블과 관련된 TableGroup, OrderTable table 패키지로 이동
