+++ src/main/java/kitchenpos/table/application/response/OrderTableResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.application.response;
+package kitchenpos.table.application.response;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 
 public class OrderTableResponse {
