+++ src/main/java/kitchenpos/table/application/response/TableGroupResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.application.response;
+package kitchenpos.table.application.response;
 
 import java.time.LocalDateTime;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 public class TableGroupResponse {
